backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyJWT
passlib[bcrypt]
bcrypt<4.0.0
PyNaCl
python-multipart
